+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This changelog follows the principles of [Keep a Changelog](https://keepachangel
 - New Use Case: [Create a Template](./docs/useCases.md#create-a-template) under Templates.
 - New Use Case: [Get a Template](./docs/useCases.md#get-a-template) under Templates.
 - New Use Case: [Delete a Template](./docs/useCases.md#delete-a-template) under Templates.
+- Templates: Added `setDefaultTemplate` use case and repository method to support Dataverse endpoint `POST /dataverses/{id}/template/default/{templateId}`.
+- Templates: Added `removeDefaultTemplate` use case and repository method to support Dataverse endpoint `DELETE /dataverses/{id}/template/default`.
 - New Use Case: [Update Terms of Access](./docs/useCases.md#update-terms-of-access).
 
 ### Changed